Simple demo drug screening for one drug (Methotrexate). The results have been generated for demonstration purpose. There is two studies present in the dataset. There is also only two types of screening in the dataset. Some entries have a value for a second drug. There is not entry for a third drug. There is no background drug. At last, there is two types of dosage.

Simple demo drug screening for one drug (Methotrexate). The results have been generated for demonstration purpose. There is two studies present in the dataset. There is also only two types of screening in the dataset. Some entries have a value for a second drug. There is not entry for a third drug. There is no background drug. At last, there is two types of dosage.

data(drugScreening)

Format

a data.frame with the following columns:

organoid_id a character string representing the organoid identifier.

plate a character string representing the plate identifier.

passage_number a character string representing the passage number of the organoid.

timestamp a character string representing the date when the screening was done.

study a character string representing the study.

screen_type a character string representing the screen type.

z_prime a numeric representing the assay quality.

drug_a a character string representing the name of the first drug used for the drug screen. There is at least one drug used in a drug screening.

drug_b a character string representing the name of the second drug used for the drug screen. When no second drug has been used, the value is 'N/A'.

drug_c a character string representing the name of the third drug used for the drug screen. When no third drug has been used, the value is 'N/A'.

drug_background a character string representing the name of the drug used as background for the drug screen. When no drug has been used as background, the value is 'N/A'.

dosage_type a character string representing the type of dosage.

dosage_concentration_a a numeric representing the dosage of the first drug used for the drug screen. This column should always have a value as there is at least one drug used in a drug screening.

dosage_concentration_b a numeric representing the dosage of the second drug used for the drug screen. When no second drug has been used, the value is 0.

dosage_concentration_c a numeric representing the dosage of the third drug used for the drug screen. When no third drug has been used, the value is 0.

dosage_concentration_background a numeric representing the dosage of the background drug used for the drug screen. When no background drug has been used, the value is 0.

percent_viability a numeric representing the percentage of viability of the organoids under the specific screening conditions.

relative_auc a numeric representing the relative AUC (area under the curve) for the specified organoids under specific conditions. The AUC is calculated using the percentage of viability under different drug concentrations. So, the same relative AUC is repeated on multiple lines.

See also

Examples

## Load drug screen dataset for 1 drug
data(drugScreening)

## Calculate the extreme organoids for the methotrexate drug screening
## using a quantile of 1/3
results <- selectOrgForOneDrug(drugScreening=drugScreening,
    drugName="Methotrexate", study="MEGA-TEST", screenType="TEST-01",
    doseType="Averaged", quantile=1/4)

## The information of the extreme organoids is found it the 'extreme' entry
results$extreme
#>   organoid_id relative_auc     group
#> 1      hT3312     46.61114 SENSITIVE
#> 2      hT1001     52.61044 SENSITIVE
#> 3      hT2211     53.57622   AVERAGE
#> 4      hT1919     56.53002   AVERAGE
#> 5      hT1051     60.28033   AVERAGE
#> 6      hT1082     68.93726 RESISTANT
#> 7      hT2212     75.67628 RESISTANT