Simple demo drug screening for one drug (Methotrexate). The results have been generated for demonstration purpose. There is two studies present in the dataset. There is also only two types of screening in the dataset. Some entries have a value for a second drug. There is not entry for a third drug. There is no background drug. At last, there is two types of dosage.

Simple demo drug screening for one drug (Methotrexate). The results have been generated for demonstration purpose. There is two studies present in the dataset. There is also only two types of screening in the dataset. Some entries have a value for a second drug. There is not entry for a third drug. There is no background drug. At last, there is two types of dosage.

data(drugScreeningMethoSet)

Format

a data.frame with the following columns:

organoid_id a character string representing the organoid identifier.

timestamp a character string representing the date when the screening was done.

study a character string representing the study.

screen_type a character string representing the screen type.

drug_a a character string representing the name of the first drug used for the drug screen. There is at least one drug used in a drug screening.

drug_b a character string representing the name of the second drug used for the drug screen. When no second drug has been used, the value is 'N/A'.

drug_c a character string representing the name of the third drug used for the drug screen. When no third drug has been used, the value is 'N/A'.

drug_background a character string representing the name of the drug used as background for the drug screen. When no drug has been used as background, the value is 'N/A'.

dosage_type a character string representing the type of dosage.

relative_auc a numeric representing the relative AUC (area under the curve) for the specified organoids under specific conditions. The AUC is calculated using the percentage of viability under different drug concentrations. So, the same relative AUC is repeated on multiple lines.

See also

Examples

## Load drug screen dataset and patient information for methotrexate
data(drugScreeningMethoSet)
data(patientInfoMethoSet)

## Retain unreplicated samples
cleanData <- selectNoReplicateOrganoids(drugScreening=drugScreeningMethoSet,
    patientInfo=patientInfoMethoSet)

## Calculate the extreme organoids for the methotrexate drug screening
## using a quantile of 1/3
results <- getClassOneDrug(drugScreening=cleanData,
    drugName="Methotrexate", study="MEGA-TEST", screenType="TEST-01",
    doseType="Averaged", quantile=1/3)

## The information of the extreme organoids is found it the 'extreme' entry
head(results$extreme)
#>   organoid_id relative_auc     group
#> 1       hT402     5.477998 SENSITIVE
#> 2       hT209    14.427644 SENSITIVE
#> 3       hT204    16.741371 SENSITIVE
#> 4       hT598    19.086306 SENSITIVE
#> 5       hT468    19.532156 SENSITIVE
#> 6       hT317    22.479160 SENSITIVE

## Fisher test on ancestry
fisherT <- fisherCategoricalVariable(drugQuantile=results,
    category="ancestry")

fisherT
#> $Fisher
#> 
#> 	Fisher's Exact Test for Count Data
#> 
#> data:  allTable
#> p-value = 0.0302
#> alternative hypothesis: two.sided
#> 
#> 
#> $table
#>           AFR EAS AMR EUR SAS
#> SENSITIVE   8   5   7  18   5
#> RESISTANT  19   6   4   7   7
#> 
#> $category
#> [1] "ancestry"
#>