Simple demo patient information dataset. The information is related to the organoids present in the 'drugScreening' dataset.

Simple demo patient information dataset. The information is related to the organoids present in the 'drugScreening' dataset.

data(patientInfo)

Format

a data.frame with the following columns:

organoid_id a character string representing the organoid identifier.

patient_id a character string representing the patient identifier.

See also

Examples

## Load drug screen dataset for 1 drug
data(drugScreening)

## Load patient information dataset for 1 drug
data(patientInfo)

## Select unrelated organoids in the drug screening
results <- selectNoReplicateOrganoids(drugScreening=drugScreening,
    patientInfo=patientInfo)

## The drug screen dataset has been filtered
results
#>        organoid_id Column1  plate passage_number timestamp      study
#> hT1001      hT1001       2 Plate1            P16  2/2/1988  MEGA-TEST
#> hT102        hT102       5 Plate1            P14  2/2/1994  MEGA-TEST
#> hT1919      hT1919       5 Plate1            P13  1/1/1970  MEGA-TEST
#> hT192        hT192       4 Plate1            P11  2/2/1994  MEGA-TEST
#> hT2211      hT2211       5 Plate1            P14  1/1/1970  MEGA-TEST
#> hT2212      hT2212       4 Plate1            P29  1/1/1970  MEGA-TEST
#> hT2213      hT2213       5 Plate2            P12  1/1/1970  MEGA-TEST
#> hT2251      hT2251       5 Plate1            P13  1/1/1970 SUPER-TEST
#> hT322        hT322       4 Plate1            P11  1/1/2022  MEGA-TEST
#> hT999        hT999       4 Plate1             P9  2/2/1994  MEGA-TEST
#>        screen_type   z_prime       drug_a drug_b drug_c drug_background
#> hT1001     TEST-01 0.2732018 Methotrexate    N/A    N/A             N/A
#> hT102      TEST-01 0.2882018 Methotrexate    N/A    N/A             N/A
#> hT1919     TEST-01 0.4589293 Methotrexate    N/A    N/A             N/A
#> hT192      TEST-01 0.2120181 Methotrexate    N/A    N/A             N/A
#> hT2211     TEST-01 0.0728753 Methotrexate    N/A    N/A             N/A
#> hT2212     TEST-01 0.0744753 Methotrexate    N/A    N/A             N/A
#> hT2213     TEST-02 0.0718753 Methotrexate    N/A    N/A             N/A
#> hT2251     TEST-01 0.2500374 Methotrexate    N/A    N/A             N/A
#> hT322      TEST-01 0.4219293 Methotrexate    N/A    N/A             N/A
#> hT999      TEST-01 0.5520181 Methotrexate    N/A    N/A             N/A
#>        dosage_type dosage_concentration_a dosage_concentration_b
#> hT1001    Averaged              0.8518696                      0
#> hT102     Averaged              0.3331993                      0
#> hT1919    Averaged             -1.8477117                      0
#> hT192     Averaged             -1.1180171                      0
#> hT2211    Averaged             -0.1481304                      0
#> hT2212    Averaged             -0.6716204                      0
#> hT2213    Averaged             -0.6446204                      0
#> hT2251    Averaged             -1.3327707                      0
#> hT322     Averaged             -1.4377117                      0
#> hT999     Averaged              1.9116352                      0
#>        dosage_concentration_c dosage_concentration_background percent_viability
#> hT1001                      0                               0          30.61149
#> hT102                       0                               0          41.10780
#> hT1919                      0                               0          71.40605
#> hT192                       0                               0          76.11993
#> hT2211                      0                               0          80.90499
#> hT2212                      0                               0          75.93872
#> hT2213                      0                               0          71.91172
#> hT2251                      0                               0          80.22052
#> hT322                       0                               0         110.33605
#> hT999                       0                               0          95.94550
#>              ic50      auc relative_auc  patient_id
#> hT1001  1.0722432 218.0053     52.61044 Patient_001
#> hT102   1.0002432 201.0053     46.11044 Patient_008
#> hT1919  1.3257419 234.2471     56.53002 Patient_002
#> hT192   0.9002432 187.0053     33.11044 Patient_009
#> hT2211  1.1417485 222.0073     53.57622 Patient_003
#> hT2212  0.8862142 246.3675     75.67628 Patient_004
#> hT2213  0.8162142 242.3675     71.67628 Patient_005
#> hT2251  1.5223589 248.8876     63.38033 Patient_006
#> hT322  14.1157419 931.1471     55.22300 Patient_007
#> hT999   0.9880232 287.0054     97.31222 Patient_010