Select the TODO

The function TODO

selectNoReplicateOrganoids(drugScreening, patientInfo)

Arguments

drugScreening

a data.frame that contains the drug screening results. The column is 'organoid_id' is mandatory.

patientInfo

a data.frame containing the meta-data information related to the organoids. The mandatory columns are: 'organoid_id' and 'patient_id'.

Value

a data.frame with the drug screening results filtered to contain with only one organoid per patient. The extra columns from the patient information dataset is added as extra columns to the final dataset.

Author

Astrid Deschênes, Pascal Belleau

Examples

## Load drug screen dataset for 1 drug
data(drugScreening)

## Load patient information dataset for 1 drug
data(patientInfo)

## Set seed to get reproducible results
set.seed(1212)

results <- selectNoReplicateOrganoids(drugScreening=drugScreening,
    patientInfo=patientInfo)

## TODO
results
#>        organoid_id Column1  plate passage_number timestamp      study
#> hT102        hT102       4 Plate1            P14  2/2/1994  MEGA-TEST
#> hT1082      hT1082       3 Plate1            P13  2/2/1988  MEGA-TEST
#> hT192        hT192       4 Plate1            P11  2/2/1994  MEGA-TEST
#> hT2211      hT2211       3 Plate1            P14  1/1/1970  MEGA-TEST
#> hT2212      hT2212       3 Plate1            P29  1/1/1970  MEGA-TEST
#> hT2213      hT2213       5 Plate2            P12  1/1/1970  MEGA-TEST
#> hT2251      hT2251       3 Plate1            P13  1/1/1970 SUPER-TEST
#> hT322        hT322       5 Plate1            P11  1/1/2022  MEGA-TEST
#> hT3312      hT3312       3 Plate1            P12  1/1/1970  MEGA-TEST
#> hT999        hT999       5 Plate1             P9  2/2/1994  MEGA-TEST
#>        screen_type   z_prime       drug_a drug_b drug_c drug_background
#> hT102      TEST-01 0.2882018 Methotrexate    N/A    N/A             N/A
#> hT1082     TEST-01 0.2362953 Methotrexate    N/A    N/A             N/A
#> hT192      TEST-01 0.2120181 Methotrexate    N/A    N/A             N/A
#> hT2211     TEST-01 0.0728753 Methotrexate    N/A    N/A             N/A
#> hT2212     TEST-01 0.0744753 Methotrexate    N/A    N/A             N/A
#> hT2213     TEST-02 0.0718753 Methotrexate    N/A    N/A             N/A
#> hT2251     TEST-01 0.2500374 Methotrexate    N/A    N/A             N/A
#> hT322      TEST-01 0.4219293 Methotrexate    N/A    N/A             N/A
#> hT3312     TEST-01 0.1628113 Methotrexate    N/A    N/A             N/A
#> hT999      TEST-01 0.5520181 Methotrexate    N/A    N/A             N/A
#>        dosage_type dosage_concentration_a dosage_concentration_b
#> hT102     Averaged             -0.1231304                      0
#> hT1082    Averaged              1.9956352                      0
#> hT192     Averaged             -1.1180171                      0
#> hT2211    Averaged             -1.3027707                      0
#> hT2212    Averaged              0.5051500                      0
#> hT2213    Averaged             -2.1132728                      0
#> hT2251    Averaged              0.6518696                      0
#> hT322     Averaged             -1.2127707                      0
#> hT3312    Averaged             -1.3027707                      0
#> hT999     Averaged              1.7446352                      0
#>        dosage_concentration_c dosage_concentration_background percent_viability
#> hT102                       0                               0          70.12445
#> hT1082                      0                               0          36.63934
#> hT192                       0                               0          76.11993
#> hT2211                      0                               0          67.49417
#> hT2212                      0                               0          46.14517
#> hT2213                      0                               0          93.53307
#> hT2251                      0                               0          43.36855
#> hT322                       0                               0          90.23416
#> hT3312                      0                               0          61.06084
#> hT999                       0                               0          95.22267
#>              ic50      auc relative_auc  patient_id
#> hT102   1.0002432 201.0053     46.11044 Patient_008
#> hT1082  3.3907678 285.6598     68.93726 Patient_001
#> hT192   0.9002432 187.0053     33.11044 Patient_009
#> hT2211  1.1417485 222.0073     53.57622 Patient_003
#> hT2212  0.8862142 246.3675     75.67628 Patient_004
#> hT2213  0.8162142 242.3675     71.67628 Patient_005
#> hT2251  1.5223589 248.8876     63.38033 Patient_006
#> hT322  14.1157419 931.1471     55.22300 Patient_007
#> hT3312  0.7672968 193.1456     46.61114 Patient_002
#> hT999   0.9880232 287.0054     97.31222 Patient_010