Validate input parameters for the selectOrgForOneDrug() function

The function validates the input parameters for the selectOrgForOneDrug() function.

validateSelectOrgForOneDrug(
  drugScreening,
  drugName,
  study,
  screenType,
  doseType,
  quantile
)

Arguments

drugScreening

a single character string representing the path and name of the drug screening file that contains the information needed to run the organoid selection. Those columns are mandatory: 'organoid_id', 'timestamp', 'study', 'screen_type', 'dosage_type', 'drug_a', 'drug_b', 'drug_c', 'drug_background' and 'relative_auc'.

drugName

a single character string representing the name of the drug selected for the analyses.

study

a vector of character string representing the name of the study selected for the analyses.

screenType

a vector of character string representing the type of screening selected for the analyses.

doseType

a single character string representing the type of dosage selected for the analyses.

quantile

a single positive numeric between 0 and 0.5 indicating the quantile used to select the organoids. Default: 1/3.

Value

the value 0L when successful.

Author

Astrid Deschênes, Pascal Belleau

Examples

## Load drug screen dataset for 1 drug
data(drugScreening)

## Validation should return OL as all parameters are valids
OrganoidsToolBox:::validateSelectOrgForOneDrug(drugScreening=drugScreening,
    drugName="Methotrexate", study="MEGA-TEST", screenType="TEST-01",
    doseType="Averaged", quantile=1/3)
#> [1] 0