Simple demo patient information dataset. The information is related to the organoids present in the 'drugScreening' dataset.

Simple demo patient information dataset. The information is related to the organoids present in the 'drugScreening' dataset.

data(drugScreening)

Format

a data.frame with the following columns:

organoid_id a character string representing the organoid identifier.

patient_id a character string representing the patient identifier.

See also

Examples

## Load drug screen dataset for 1 drug
data(drugScreening)

## Load patient information dataset for 1 drug
data(patientInfo)

## Select unrelated organoids in the drug screening
results <- selectNoReplicateOrganoids(drugScreening=drugScreening,
    patientInfo=patientInfo)

## The drug screen dataset has been filtered
results
#>        organoid_id Column1  plate passage_number timestamp      study
#> hT102        hT102       4 Plate1            P14  2/2/1994  MEGA-TEST
#> hT1051      hT1051       5 Plate1            P11  1/1/1970  MEGA-TEST
#> hT1082      hT1082       5 Plate1            P13  2/2/1988  MEGA-TEST
#> hT192        hT192       4 Plate1            P11  2/2/1994  MEGA-TEST
#> hT2211      hT2211       4 Plate1            P14  1/1/1970  MEGA-TEST
#> hT2212      hT2212       5 Plate1            P29  1/1/1970  MEGA-TEST
#> hT2213      hT2213       4 Plate2            P12  1/1/1970  MEGA-TEST
#> hT2251      hT2251       4 Plate1            P13  1/1/1970 SUPER-TEST
#> hT322        hT322       5 Plate1            P11  1/1/2022  MEGA-TEST
#> hT999        hT999       3 Plate1             P9  2/2/1994  MEGA-TEST
#>        screen_type   z_prime       drug_a   drug_b drug_c drug_background
#> hT102      TEST-01 0.2882018 Methotrexate      N/A    N/A             N/A
#> hT1051     TEST-01 0.2500374 Methotrexate      N/A    N/A             N/A
#> hT1082     TEST-01 0.2362953 Methotrexate      N/A    N/A             N/A
#> hT192      TEST-01 0.2120181 Methotrexate      N/A    N/A             N/A
#> hT2211     TEST-01 0.0728753 Methotrexate      N/A    N/A             N/A
#> hT2212     TEST-01 0.0733353 Methotrexate Abraxane    N/A             N/A
#> hT2213     TEST-02 0.0718753 Methotrexate      N/A    N/A             N/A
#> hT2251     TEST-01 0.2500374 Methotrexate      N/A    N/A             N/A
#> hT322      TEST-01 0.4219293 Methotrexate      N/A    N/A             N/A
#> hT999      TEST-01 0.5520181 Methotrexate      N/A    N/A             N/A
#>        dosage_type dosage_concentration_a dosage_concentration_b
#> hT102     Averaged             -1.8017117                      0
#> hT1051    Averaged              1.4533183                      0
#> hT1082    Averaged             -0.1481304                      0
#> hT192     Averaged              1.4463183                      0
#> hT2211    Averaged             -0.7495800                      0
#> hT2212    Averaged             -1.4336204                      0
#> hT2213    Averaged              0.6066500                      0
#> hT2251    Averaged             -1.8277117                      0
#> hT322     Averaged              0.3211993                      0
#> hT999     Averaged             -0.1231304                      0
#>        dosage_concentration_c dosage_concentration_background percent_viability
#> hT102                       0                               0         81.119934
#> hT1051                      0                               0         25.571997
#> hT1082                      0                               0         74.661209
#> hT192                       0                               0         11.945501
#> hT2211                      0                               0         79.552841
#> hT2212                      0                               0          8.727732
#> hT2213                      0                               0         42.115171
#> hT2251                      0                               0         84.427732
#> hT322                       0                               0         92.081015
#> hT999                       0                               0         97.664067
#>              ic50       auc relative_auc  patient_id
#> hT102   1.0002432 201.00535    46.110445 Patient_008
#> hT1051  1.7733589 249.78756    60.280332 Patient_002
#> hT1082  3.3907678 285.65983    68.937256 Patient_001
#> hT192   0.9002432 187.00535    33.110445 Patient_009
#> hT2211  1.1417485 222.00729    53.576219 Patient_003
#> hT2212  0.7444589  49.99956     6.330331 Patient_004
#> hT2213  0.8162142 242.36746    71.676280 Patient_005
#> hT2251  1.5223589 248.88756    63.380332 Patient_006
#> hT322  14.1157419 931.14714    55.223002 Patient_007
#> hT999   0.9880232 287.00535    97.312221 Patient_010